додавання та вивід фото на сайті зроблено

+++ src/main/java/ua/nure/shoestore/entity/Shoe.java
@@ -16,5 +16,5 @@ public class Shoe {
     private BigDecimal price;
     private String name;
     private int amount;
-    private int imageId;
+    private Long imageId;
 }
